feat(cli): unify brew commands under single interface for flexibility

+++ dot_dotenv/commands.zsh
@@ -1,14 +1,18 @@
-function brew_dump() {
-  # this commands dumps all the brew packages installed in a Brewfile
-  # this requires that ozxide is installed
-  local chezmoi_dir=$(chezmoi source-path)
-  local current_dir=$(pwd)
-  cd $chezmoi_dir
-  cd dot_brew
-  rm -rf Brewfile
-  brew bundle dump
-  sed  -i '' '/^vscode/d' Brewfile
-  cd $current_dir
+function brew() { 
+  if [ "$1" = "dump" ]; then
+    # this commands dumps all the brew packages installed in a Brewfile
+    # this requires that ozxide is installed
+    local chezmoi_dir=$(chezmoi source-path)
+    local current_dir=$(pwd)
+    cd $chezmoi_dir
+    cd dot_brew
+    rm -rf Brewfile
+    brew bundle dump
+    sed  -i '' '/^vscode/d' Brewfile
+    cd $current_dir
+  else 
+    command brew "$@"
+  fi
 }
 
 function renew_tokens() {
@@ -44,5 +48,13 @@ function update() {
   brew upgrade
   brew cleanup
   brew doctor
-  brew_dump
+  brew dump
+}
+
+function git() {
+  if [ "$1" = "cleanup" ]; then
+    git fetch -p && git branch -vv | grep ': gone' | awk '{print $1}' | xargs git branch -D
+  else
+    command git "$@"
+  fi
 }
